+++ src/components/post.tsx
@@ -1,4 +1,4 @@
-import { PostDTO } from './types/dto'
+import { PostDTO } from '../types/dto'
 import classes from './Post.module.css'
 import { useState } from 'react'
